støtter automatisk ferdigstilling av journalposter
