+++ src/test/kotlin/no/nav/tiltakspenger/soknad/api/joark/JoarkClientTest.kt
@@ -23,34 +23,7 @@ internal class JoarkClientTest {
     private val baseurl = "http://dokarkiv"
 
     @Test
-    fun `joark svarer alt ok`() {
-        val mock = MockEngine {
-            respond(
-                content = okSvarJoark,
-                status = HttpStatusCode.Created,
-                headers = headersOf(HttpHeaders.ContentType, ContentType.Application.Json.toString()),
-            )
-        }
-
-        val client = httpClientGeneric(mock)
-        val joarkClient = JoarkClient(
-            client = client,
-            baseUrl = baseurl,
-        ) { getMockToken() }
-
-        runTest {
-            val resp = joarkClient.opprettJournalpost(
-                dokumentInnhold = dokument,
-                søknadId = søknadId,
-                callId = "123",
-            )
-
-            resp shouldBe journalpostId
-        }
-    }
-
-    @Test
-    fun `joark svarer ikke ferdigstilt`() {
+    fun `opprettJournalpost - skal ikke ferdigstille, blir opprettet - returnerer journalpostid`() {
         val mock = MockEngine {
             respond(
                 content = svarIkkeFerdigstiltJoark,
@@ -59,12 +32,6 @@ internal class JoarkClientTest {
             )
         }
 
-//        val joarkResonse = JoarkClient.JoarkResponse(
-//            journalpostId = journalpostId,
-//            journalpostferdigstilt = false,
-//            dokumenter = listOf(dokumentResponse),
-//        )
-
         val client = httpClientGeneric(mock)
         val joarkClient = JoarkClient(
             client = client,
@@ -73,7 +40,7 @@ internal class JoarkClientTest {
 
         runTest {
             val resp = joarkClient.opprettJournalpost(
-                dokumentInnhold = dokument,
+                request = journalpostRequest(),
                 søknadId = søknadId,
                 callId = "123",
             )
@@ -83,10 +50,64 @@ internal class JoarkClientTest {
     }
 
     @Test
+    fun `opprettJournalpost - skal ferdigstille, blir opprettet og ferdigstilt - returnerer journalpostid`() {
+        val mock = MockEngine {
+            respond(
+                content = svarFerdigstiltJoark,
+                status = HttpStatusCode.Created,
+                headers = headersOf(HttpHeaders.ContentType, ContentType.Application.Json.toString()),
+            )
+        }
+
+        val client = httpClientGeneric(mock)
+        val joarkClient = JoarkClient(
+            client = client,
+            baseUrl = baseurl,
+        ) { getMockToken() }
+
+        runTest {
+            val resp = joarkClient.opprettJournalpost(
+                request = journalpostRequest("1234", "saksnummer"),
+                søknadId = søknadId,
+                callId = "123",
+            )
+
+            resp shouldBe journalpostId
+        }
+    }
+
+    @Test
+    fun `opprettJournalpost - skal ferdigstille, blir opprettet men ikke ferdigstilt - kaster feil`() {
+        val mock = MockEngine {
+            respond(
+                content = svarIkkeFerdigstiltJoark,
+                status = HttpStatusCode.Created,
+                headers = headersOf(HttpHeaders.ContentType, ContentType.Application.Json.toString()),
+            )
+        }
+
+        val client = httpClientGeneric(mock)
+        val joarkClient = JoarkClient(
+            client = client,
+            baseUrl = baseurl,
+        ) { getMockToken() }
+
+        runTest {
+            shouldThrow<RuntimeException> {
+                joarkClient.opprettJournalpost(
+                    request = journalpostRequest("1234", "saksnummer"),
+                    søknadId = søknadId,
+                    callId = "123",
+                )
+            }
+        }
+    }
+
+    @Test
     fun `hvis joark svarer med 409 Conflict returnerer opprettJournalpost en journalpostId hvis vi har fått en`() {
         val mock = MockEngine {
             respond(
-                content = okSvarJoark,
+                content = svarIkkeFerdigstiltJoark,
                 status = HttpStatusCode.Conflict,
                 headers = headersOf(HttpHeaders.ContentType, ContentType.Application.Json.toString()),
             )
@@ -100,7 +121,7 @@ internal class JoarkClientTest {
 
         runTest {
             val resp = joarkClient.opprettJournalpost(
-                dokumentInnhold = dokument,
+                request = journalpostRequest(),
                 søknadId = søknadId,
                 callId = "123",
             )
@@ -127,7 +148,7 @@ internal class JoarkClientTest {
         runTest {
             shouldThrow<RuntimeException> {
                 joarkClient.opprettJournalpost(
-                    dokumentInnhold = dokument,
+                    request = journalpostRequest(),
                     søknadId = søknadId,
                     callId = "123",
                 )
@@ -135,77 +156,14 @@ internal class JoarkClientTest {
         }
     }
 
-    @Test
-    fun `joark svarer med null i ferdigstilt`() {
-        val mock = MockEngine {
-            respond(
-                content = svarNullIFerdigstiltJoark,
-                status = HttpStatusCode.Created,
-                headers = headersOf(HttpHeaders.ContentType, ContentType.Application.Json.toString()),
-            )
-        }
-
-//        val joarkResonse = JoarkClient.JoarkResponse(
-//            journalpostId = journalpostId,
-//            journalpostferdigstilt = null,
-//            dokumenter = listOf(dokumentResponse),
-//        )
-
-        val client = httpClientGeneric(mock)
-        val joarkClient = JoarkClient(
-            client = client,
-            baseUrl = baseurl,
-        ) { getMockToken() }
-
-        runTest {
-            val resp = joarkClient.opprettJournalpost(
-                dokumentInnhold = dokument,
-                søknadId = søknadId,
-                callId = "123",
-            )
-
-            resp shouldBe journalpostId
-        }
-    }
-
-    @Test
-    fun `joark svarer uten journalpostid`() {
-        val mock = MockEngine {
-            respond(
-                content = svarUtenJournalpostId,
-                status = HttpStatusCode.Created,
-                headers = headersOf(HttpHeaders.ContentType, ContentType.Application.Json.toString()),
-            )
-        }
-
-        val joarkResonse = JoarkClient.JoarkResponse(
-            journalpostId = null,
-            journalpostferdigstilt = null,
-            dokumenter = listOf(dokumentResponse),
-        )
-
-        val client = httpClientGeneric(mock)
-        val joarkClient = JoarkClient(
-            client = client,
-            baseUrl = baseurl,
-        ) { getMockToken() }
-
-        runTest {
-            shouldThrow<IllegalStateException> {
-                joarkClient.opprettJournalpost(
-                    dokumentInnhold = dokument,
-                    søknadId = søknadId,
-                    callId = "123",
-                )
-            }.message shouldBe "Fikk 201 Created fra Joark, men vi fikk ingen journalpostId. response=JoarkResponse(journalpostId=null, journalpostferdigstilt=null, dokumenter=[Dokumenter(dokumentInfoId=485227498, tittel=Søknad om tiltakspenger)])"
-        }
-    }
-
-    private suspend fun getMockToken(): AccessToken {
+    private fun getMockToken(): AccessToken {
         return AccessToken("token", Instant.now().plusSeconds(3600)) {}
     }
 
-    private val dokument = Journalpost.Søknadspost.from(
+    private fun journalpostRequest(
+        journalforendeEnhet: String? = null,
+        saksnummer: String? = null,
+    ) = JournalpostRequest.from(
         fnr = "ident",
         søknad = søknad(),
         pdf = "dette er pdf innholdet".toByteArray(),
@@ -216,70 +174,21 @@ internal class JoarkClientTest {
                 dokument = "vedleggInnhold".toByteArray(),
             ),
         ),
+        journalforendeEnhet = journalforendeEnhet,
+        saksnummer = saksnummer,
     )
 
-    private val dokumentInfoId = "485227498"
-    private val dokumentInfoVedleggId = "485227499"
-    private val dokumentTittel = "Søknad om tiltakspenger"
-    private val dokumentVedleggFilnavn = "filnavnVedlegg"
-    private val dokumentResponse = JoarkClient.Dokumenter(
-        dokumentInfoId = dokumentInfoId,
-        tittel = dokumentTittel,
-
-    )
-    private val okSvarJoark = """
+    private val svarFerdigstiltJoark = """
         {
           "journalpostId": "$journalpostId",
-          "journalpostferdigstilt": true,
-          "dokumenter": [
-            {
-              "dokumentInfoId": "$dokumentInfoId",
-              "tittel": "$dokumentTittel"
-            },
-            {
-              "dokumentInfoId": "$dokumentInfoVedleggId",
-              "tittel": "$dokumentVedleggFilnavn"
-            }
-          ]
+          "journalpostferdigstilt": true
         }
     """.trimIndent()
 
     private val svarIkkeFerdigstiltJoark = """
         {
           "journalpostId": "$journalpostId",
-          "journalpostferdigstilt": false,
-          "dokumenter": [
-            {
-              "dokumentInfoId": "$dokumentInfoId",
-              "tittel": "$dokumentTittel"
-            }
-          ]
-        }
-    """.trimIndent()
-
-    private val svarNullIFerdigstiltJoark = """
-        {
-          "journalpostId": "$journalpostId",
-          "journalpostferdigstilt": null,
-          "dokumenter": [
-            {
-              "dokumentInfoId": "$dokumentInfoId",
-              "tittel": "$dokumentTittel"
-            }
-          ]
-        }
-    """.trimIndent()
-
-    private val svarUtenJournalpostId = """
-        {
-          "journalpostId": null,
-          "journalpostferdigstilt": null,
-          "dokumenter": [
-            {
-              "dokumentInfoId": "$dokumentInfoId",
-              "tittel": "$dokumentTittel"
-            }
-          ]
+          "journalpostferdigstilt": false
         }
     """.trimIndent()
 }
