+++ src/test/kotlin/no/nav/tiltakspenger/soknad/api/soknad/MottattSøknadServiceTest.kt
@@ -37,7 +37,7 @@ internal class MottattSøknadServiceTest {
         coEvery { mock.lagPdf(any()) }
     }
     private val mockJoarkService = mockk<JoarkService>().also { mock ->
-        coEvery { mock.sendPdfTilJoark(any(), any(), any(), any(), any(), any()) }
+        coEvery { mock.sendPdfTilJoark(any(), any(), any(), any(), any(), any(), any(), any()) }
     }
 
     private val gyldigSpørsmålsbesvarelser = spørsmålsbesvarelser()
